Use mixin instead of helper class
