+++ types/foundry/client/applications/sheets/_module.d.mts
@@ -1,6 +1,7 @@
 export { default as ItemSheetV2 } from "./item-sheet.mjs";
 export * from "./scene-config.mjs";
 export { default as SceneConfig } from "./scene-config.mjs";
+export { default as TokenApplicationMixin } from "./token/mixin.mts";
 export { default as PrototypeTokenConfig } from "./token/prototype-config.mjs";
 export { default as TokenConfig } from "./token/token-config.mjs";
 export * from "./user-config.mjs";
